Refactor FilePath into different package
