+++ src/main/java/seedu/address/model/FilePath.java
@@ -1,6 +1,6 @@
 //@@author LeowWB
 
-package seedu.address.model.util;
+package seedu.address.model;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
