+++ src/test/java/seedu/address/logic/commands/ExportCommandTest.java
@@ -22,7 +22,7 @@ import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.category.Category;
-import seedu.address.model.util.FilePath;
+import seedu.address.model.FilePath;
 
 /**
  * Contains integration tests (interaction with the Model) for {@code FindCommand}.
